+++ src/main/java/com/pnu/momeet/domain/chatting/enums/ChatActionType.java
@@ -7,6 +7,7 @@ public enum ChatActionType {
     JOIN,       // 모임 참여
     KICKED,     // 모임 강퇴
     EXIT,       // 모임 나가기
+    STARTED,      // 모임 시작(OPEN -> IN_PROGRESS)
     FINISH,     // 모임 종료
     CANCELED,     // 모임 취소
     CANCELED_BY_ADMIN // 모임 관리자에 의한 모임 취소
